requirements.txt
jinja2==3.0.3
prettytable==2.4.0
pyyaml==6.0